feat(kimi): ship data preparation and LoRA conversion CLIs
Move the existing adapter converter into an installable tools package without changing conversion math. Add deterministic NekoQA preparation and candidate save/resume instructions; keep packing disabled.
