+++ kt-kernel/python_tools/__init__.py
@@ -1,0 +1,1 @@
+"""Offline KT preparation tools; importing these does not load native kernels."""
